requirements.txt
mlflow<3,>=2.3
cffi==1.15.1
cloudpickle==2.2.1
defusedxml==0.7.1
imbalanced-learn==0.10.1
lightgbm==3.3.5
matplotlib==3.7.1
numpy==1.23.5
packaging==23.1
pandas==2.0.1
psutil==5.9.5
scikit-learn==1.2.2
scipy==1.10.1
shap~=0.41.0
fastapi~=0.98.0
uvicorn~=0.22.0
streamlit~=1.24.0
requests~=2.31.0
seaborn~=0.12.2
Pillow~=9.5.0